HW9/hw9-docker/requirements.txt
flask
google-cloud-storage
google-cloud-pubsub
google-cloud-logging
flask_cors
google-cloud-compute
beautifulsoup4